Map over each combination of list elements simultaneously via futures

These functions work exactly the same as xmap() functions, but allow you to run the map in parallel using future::future()

future_xmap(.l, .f, ..., .progress = FALSE, .options = furrr::furrr_options())

future_xmap_chr(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_dbl(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_dfc(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_dfr(
  .l,
  .f,
  ...,
  .id = NULL,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_int(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_lgl(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_raw(
  .l,
  .f,
  ...,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xwalk(.l, .f, ..., .progress = FALSE, .options = furrr::furrr_options())

Arguments

.l
A list of vectors, such as a data frame. The length of .l determines the number of arguments that .f will be called with. List names will be used if present.

.f
A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans

.options
The future specific options to use with the workers. This must be the result from a call to furrr::furrr_options()

.id
Either a string or NULL. If a string, the output will contain a variable with that name, storing either the name (if .x is named) or the index (if .x is unnamed) of the input. If NULL, the default, no variable will be created.

Only applies to _dfr variant.

Value

An atomic vector, list, or data frame, depending on the suffix. Atomic vectors and lists will be named if the first element of .l is named.

If all input is length 0, the output will be length 0. If any input is length 1, it will be recycled to the length of the longest.

See also

Examples

future_xmap(list(1:5, 1:5), ~ .y * .x)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

future_xmap_dbl(list(1:5, 1:5), ~ .y * .x)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

future_xmap_chr(list(1:5, 1:5), ~ paste(.y, "*", .x, "=", .y * .x))

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

apples_and_bananas <- list(
  x = c("apples", "bananas"),
  pattern = "a",
  replacement = c("oo", "ee")
)

future_xmap_chr(apples_and_bananas, gsub)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

formulas <- list(mpg ~ wt, mpg ~ hp)
subsets  <- split(mtcars, mtcars$cyl)

future_xmap(list(subsets, formulas), ~ lm(.y, data = .x))

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively