Parallelized cross map returning a matrix or array

Parallelized cross map returning a matrix or array

future_xmap_mat(
  .l,
  .f,
  ...,
  .names = TRUE,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

future_xmap_arr(
  .l,
  .f,
  ...,
  .names = TRUE,
  .progress = FALSE,
  .options = furrr::furrr_options()
)

Arguments

.l
A list of vectors, such as a data frame. The length of .l determines the number of arguments that .f will be called with. List names will be used if present.

.f
A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f

.names
A logical indicating whether to give names to the dimensions of the matrix or array. If inputs are named, the names are used. If inputs are unnamed, the elements of the input are used as names. Defaults to TRUE.

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans

.options
The future specific options to use with the workers. This must be the result from a call to furrr::furrr_options()

Value

A matrix (for future_xmap_mat()) or array (for future_xmap_arr()) with dimensions matching the lengths of each input in .l.

See also

Examples

future_xmap_mat(list(1:3, 1:3),  ~ ..1 * ..2)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

fruits <- c(a = "apple", b = "banana", c = "cantaloupe")
future_xmap_mat(list(1:3, fruits), paste)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

future_xmap_mat(list(1:3, fruits), paste, .names = FALSE)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively

future_xmap_arr(list(1:3, 1:3, 1:3),  ~ ..1 * ..2 * ..3)

#> Error in utils::menu(c("Multicore (recommended for non-Windows)", "Multisession (recommended for Windows)",     "Cancel"), title = paste(paste(code(fn), "is not set up to run background processes."),     paste("Please choose a", code("future"), "plan."), paste("Check",         code("help(plan, future)"), "for more details."), sep = "\n")): menu() cannot be used non-interactively